extracted the duplicated API key approval/denial logic into a shared utility function
